+++ README.md
@@ -1,11 +1,23 @@
 # vep-sim
 create randomized simulations of variant effect predictions from an input vcf file
 
+## usage
+Given an input vcf file, Using Ensembl VEP (variant effect predictor) software to predict 
+
+CREATE RANDOMIZED SIMULATIONS OF VARIANT EFFECT PREDICTIONS FROM AN INPUT VCF FILE
+
 ## inputs 
 1. VCF file
 2. BED file
 3. Fasta file
 
+
 ## outputs
 1. VEP report file
 2. statistics based off selection
+
+## installation
+```bash
+pip install vep-sim
+```
+
